requirements.txt
streamlit
langchain==0.1.14
langchain-community==0.0.30
transformers==4.39.3
sentence-transformers
huggingface_hub
sentencepiece
python-dotenv
accelerate
unstructured
numpy==1.26.4

